Refactor : 불필요한 enum 타입 제거 및 카카오 메시지 api 폴더 위치 변경 #157
