--- src/main/java/com/sparta/finalproject/domain/child/entity/ScheduleType.java
@@ -1,5 +1,0 @@
-package com.sparta.finalproject.domain.child.entity;
-
-public enum ScheduleType {
-    ENTER, EXIT
-}
