--- src/main/java/com/sparta/finalproject/global/enumType/AbsentReason.java
@@ -1,4 +1,0 @@
-package com.sparta.finalproject.global.enumType;
-
-public enum AbsentReason {
-}
